+++ src/main/java/tk/estecka/backburner/BacklogCommands.java
@@ -13,8 +13,9 @@ import net.minecraft.text.MutableText;
 import net.minecraft.text.Text;
 import net.minecraft.util.Formatting;
 import net.minecraft.util.Identifier;
-import static net.fabricmc.fabric.api.client.command.v2.ClientCommandManager.literal;
+import tk.estecka.backburner.hud.BacklogHud;
 import java.util.concurrent.CompletableFuture;
+import static net.fabricmc.fabric.api.client.command.v2.ClientCommandManager.literal;
 import static net.fabricmc.fabric.api.client.command.v2.ClientCommandManager.argument;
 import static com.mojang.brigadier.arguments.BoolArgumentType.bool;
 import static com.mojang.brigadier.arguments.BoolArgumentType.getBool;
