+++ src/main/java/tk/estecka/backburner/mixin/InGameHudMixin.java
@@ -3,7 +3,7 @@ package tk.estecka.backburner.mixin;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.gui.hud.InGameHud;
 import net.minecraft.client.util.math.MatrixStack;
-import tk.estecka.backburner.BacklogHud;
+import tk.estecka.backburner.hud.BacklogHud;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
